requirements.txt
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
xgboost==2.1.2
torch==2.5.1
networkx==3.4.2
shap==0.46.0
streamlit==1.40.1
matplotlib==3.9.2
seaborn==0.13.2
python-dotenv==1.0.1
tqdm==4.67.0
pytest==8.3.3
